backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib
bcrypt==3.2.2
numpy<2
tensorflow
pillow
geopy
email-validator
